# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de Jajurin Y Laizerovich - Provincias",
    "description": "API REST CRUD de provincias argentinas utilizando datos en memoria.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/province/": {
      "get": {
        "tags": [
          "Provincias"
        ],
        "summary": "Obtiene todas las provincias",
        "description": "Devuelve el listado completo de provincias argentinas cargadas en el sistema, sin aplicar filtros ni un orden particular.",
        "responses": {
          "200": {
            "description": "Lista de todas las provincias",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "array"
                },
                "items": {
                  "$ref": "#/definitions/Provincia"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "Error interno del servidor."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Provincias"
        ],
        "summary": "Crea una nueva provincia",
        "description": "Crea una nueva provincia a partir de los datos recibidos en el body. El ID se genera automáticamente. Antes de crearla se validan los campos obligatorios: name (mínimo 3 caracteres), full_name (mínimo 5 caracteres), latitude y longitude (numéricos) y display_order (entero, opcional).",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Datos de la provincia a crear",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProvinciaInput"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Provincia creada exitosamente",
            "schema": {
              "$ref": "#/definitions/Provincia"
            }
          },
          "400": {
            "description": "Datos inválidos. Puede deberse a: nombre obligatorio o con menos de 3 caracteres, nombre completo obligatorio o con menos de 5 caracteres, latitud/longitud obligatorias y numéricas, o display_order que no sea un número entero.",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "El nombre es obligatorio!!!"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "Error interno del servidor."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Provincias"
        ],
        "summary": "Actualiza una provincia",
        "description": "Actualiza los datos de una provincia existente. El ID enviado en el body debe corresponder a una provincia ya cargada; de lo contrario devuelve 404. Se aplican las mismas validaciones de campos que en la creación.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Datos completos de la provincia a actualizar. El ID debe corresponder a una provincia existente.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Provincia"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provincia actualizada exitosamente",
            "schema": {
              "$ref": "#/definitions/Provincia"
            }
          },
          "400": {
            "description": "Datos inválidos. Puede deberse a: ID faltante, nombre obligatorio o con menos de 3 caracteres, nombre completo obligatorio o con menos de 5 caracteres, latitud/longitud obligatorias y numéricas, o display_order que no sea un número entero.",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "La latitud debe ser un número"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Provincia no encontrada",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "Provincia no encontrada."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "Error interno del servidor."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/province/search": {
      "get": {
        "tags": [
          "Provincias"
        ],
        "summary": "Busca provincias por nombre",
        "description": "Filtra el listado de provincias devolviendo únicamente aquellas cuyo nombre contenga el texto recibido por query string (búsqueda parcial, sin distinguir mayúsculas/minúsculas).",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Texto a buscar dentro del nombre de la provincia",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Provincias que coinciden con la búsqueda",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "array"
                },
                "items": {
                  "$ref": "#/definitions/Provincia"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Falta el parámetro name",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "El parámetro \"name\" es requerido."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No se encontraron provincias que coincidan con el nombre buscado",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "No se encontraron provincias con ese nombre."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "Error interno del servidor."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/province/order": {
      "get": {
        "tags": [
          "Provincias"
        ],
        "summary": "Obtiene todas las provincias ordenadas por display_order",
        "description": "Devuelve el listado completo de provincias ordenado según el campo display_order. Si no se especifica el parámetro sort, se ordena de forma ascendente por defecto. Los únicos valores válidos para sort son \"asc\" y \"desc\".",
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "description": "Dirección del ordenamiento (asc o desc). Por defecto asc.",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Provincias ordenadas",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "array"
                },
                "items": {
                  "$ref": "#/definitions/Provincia"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "El parámetro sort tiene un valor inválido (solo se acepta \"asc\" o \"desc\")",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "El parámetro \"sort\" debe ser \"asc\" o \"desc\"."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "Error interno del servidor."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/province/{id}": {
      "get": {
        "tags": [
          "Provincias"
        ],
        "summary": "Obtiene una provincia por ID",
        "description": "Busca y devuelve una única provincia a partir de su ID numérico. Si no existe una provincia con ese ID, devuelve un error 404.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID numérico de la provincia"
          }
        ],
        "responses": {
          "200": {
            "description": "Provincia encontrada",
            "schema": {
              "$ref": "#/definitions/Provincia"
            }
          },
          "400": {
            "description": "El ID enviado no es un número válido",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "El ID debe ser un número."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Provincia no encontrada",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "Provincia no encontrada."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "Error interno del servidor."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Provincias"
        ],
        "summary": "Elimina una provincia por ID",
        "description": "Elimina definitivamente una provincia a partir de su ID numérico. Si no existe una provincia con ese ID, devuelve un error 404.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID numérico de la provincia"
          }
        ],
        "responses": {
          "200": {
            "description": "Provincia eliminada exitosamente",
            "schema": {
              "$ref": "#/definitions/Provincia"
            }
          },
          "400": {
            "description": "El ID enviado no es un número válido",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "El ID debe ser un número."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Provincia no encontrada",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "Provincia no encontrada."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "Error interno del servidor."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Provincia": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "id": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "number"
                },
                "example": {
                  "type": "number",
                  "example": 1
                },
                "description": {
                  "type": "string",
                  "example": "Identificador único de la provincia, generado automáticamente por el sistema."
                }
              }
            },
            "name": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "Buenos Aires"
                },
                "description": {
                  "type": "string",
                  "example": "Nombre corto o de uso común de la provincia."
                }
              }
            },
            "full_name": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "Provincia de Buenos Aires"
                },
                "description": {
                  "type": "string",
                  "example": "Nombre oficial completo de la provincia."
                }
              }
            },
            "latitude": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "-36.6769"
                },
                "description": {
                  "type": "string",
                  "example": "Latitud geográfica del centro de la provincia. Se almacena como texto para conservar el formato numérico exacto (incluye signo y decimales)."
                }
              }
            },
            "longitude": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "-60.5588"
                },
                "description": {
                  "type": "string",
                  "example": "Longitud geográfica del centro de la provincia. Se almacena como texto por el mismo motivo que la latitud."
                }
              }
            },
            "display_order": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "number"
                },
                "example": {
                  "type": "number",
                  "example": 1
                },
                "description": {
                  "type": "string",
                  "example": "Orden numérico utilizado para mostrar las provincias en listados (por ejemplo, de norte a sur). Es opcional."
                }
              }
            }
          }
        }
      }
    },
    "ProvinciaInput": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "name": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "Buenos Aires"
                },
                "description": {
                  "type": "string",
                  "example": "Nombre corto o de uso común de la provincia."
                }
              }
            },
            "full_name": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "Provincia de Buenos Aires"
                },
                "description": {
                  "type": "string",
                  "example": "Nombre oficial completo de la provincia."
                }
              }
            },
            "latitude": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "-36.6769"
                },
                "description": {
                  "type": "string",
                  "example": "Latitud geográfica del centro de la provincia (numérica, en formato texto)."
                }
              }
            },
            "longitude": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "-60.5588"
                },
                "description": {
                  "type": "string",
                  "example": "Longitud geográfica del centro de la provincia (numérica, en formato texto)."
                }
              }
            },
            "display_order": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "number"
                },
                "example": {
                  "type": "number",
                  "example": 1
                },
                "description": {
                  "type": "string",
                  "example": "Orden numérico opcional utilizado para mostrar las provincias en listados."
                }
              }
            }
          }
        }
      }
    }
  }
}